Backend completo SIN CORS

+++ backend/controllers/tareasControllers.js
@@ -32,8 +32,14 @@ const updateTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-    res.status(200).json(tareaUpdated)
+    //verificar que la tarea pertenezca al usuario del token
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+        res.status(200).json(tareaUpdated)
+    }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -45,9 +51,14 @@ const deleteTareas = asyncHandler(async (req, res) => {
         throw new Error('Esa tarea no existe')
     }
 
-    tarea.deleteOne()
-
-    res.status(200).json({ id: req.params.id })
+    //verificar que la tarea pertenezca al usuario del token
+    if (tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        tarea.deleteOne()
+        res.status(200).json({ id: req.params.id })
+    }
 })
 
 module.exports = {
